inclusion of a new warndialog in case the robot is not stopped
In case the robot is not stopped and an import operation, at the settings menu, is desired to be done. The .mod files cannot be uploaded. So a new warning pop up window is implemented.

Also there were some issues with the import function, as it was not entirely compliant with a RESTful app.

The previous code has been comented, just in case something appears. Once its functionality is probed it can be removed.